requirements.txt
django
django-rest-framework
requests
requests[security]
jinja2